requirements.txt
aiofiles==0.6.0
anyio==3.6.1
biopython==1.79
beautifulsoup4==4.11.1
boto3==1.26.65
celery==5.2.7
Cerberus==1.3.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.0
flower==1.2.0
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
httpx
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
orjson==3.8.0
pandas==1.5.0
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.4
PyYAML==6.0
psycopg2==2.9.5
redis==4.3.4
requests==2.28.1
retry==0.9.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
streaming-form-data
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
xmltodict==0.13.0
